+++ app/src/main/java/org/stepik/android/domain/profile/interactor/ProfileInteractor.kt
@@ -4,12 +4,12 @@ import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Single
 import org.stepic.droid.analytic.Analytic
-import org.stepic.droid.features.achievements.repository.AchievementsRepository
 import org.stepic.droid.model.AchievementFlatItem
 import org.stepic.droid.preferences.SharedPreferenceHelper
 import org.stepic.droid.ui.util.TimeIntervalUtil
 import org.stepic.droid.util.first
 import org.stepic.droid.util.toMaybe
+import org.stepik.android.domain.achievements.repository.AchievementsRepository
 import org.stepik.android.domain.profile.repository.ProfileRepository
 import org.stepik.android.domain.user.repository.UserRepository
 import org.stepik.android.domain.user_activity.repository.UserActivityRepository
