+++ app/src/main/java/org/stepik/android/view/injection/profile/ProfileComponent.kt
@@ -3,7 +3,7 @@ package org.stepik.android.view.injection.profile
 import dagger.Subcomponent
 import org.stepik.android.view.injection.user.UserDataModule
 import org.stepik.android.view.injection.user_activity.UserActivityDataModule
-import org.stepik.android.view.profile.ui.fragment.ProfileFragment
+import org.stepik.android.view.profile.ui.fragment.ProfileFragmentOld
 
 @Subcomponent(modules = [
     ProfileModule::class,
@@ -17,5 +17,6 @@ interface ProfileComponent {
         fun build(): ProfileComponent
     }
 
-    fun inject(profileFragment: ProfileFragment)
+    fun inject(profileFragmentOld: ProfileFragmentOld)
+    fun inject(profileFragment: org.stepik.android.view.profile.ui.fragment.ProfileFragment)
 }
