+++ app/src/main/java/org/stepik/android/view/profile/ui/activity/ProfileActivityOld.kt
@@ -12,8 +12,9 @@ import org.stepic.droid.base.SingleFragmentActivity
 import org.stepic.droid.ui.activities.contracts.CloseButtonInToolbar
 import org.stepic.droid.util.AppConstants
 import org.stepik.android.view.profile.ui.fragment.ProfileFragment
+import org.stepik.android.view.profile.ui.fragment.ProfileFragmentOld
 
-class ProfileActivity : SingleFragmentActivity(),
+class ProfileActivityOld : SingleFragmentActivity(),
     CloseButtonInToolbar {
 
     companion object {
@@ -40,11 +41,11 @@ class ProfileActivity : SingleFragmentActivity(),
 
         val userIdInternal: Long? = intent?.extras?.getLong(optionalUserIdKey)
         return if (userIdInternal != null && userIdInternal != 0L) {
-            ProfileFragment.newInstance(userIdInternal)
+            ProfileFragment.newInstance()//(userIdInternal)
         } else {
             val dataUri = intent?.data
             val userId = getUserId(dataUri)
-            ProfileFragment.newInstance(userId)
+            ProfileFragment.newInstance()//(userId)
         }
     }
